chore: don't bind an injection context to an injector
While binding the context to an injector seems like a good idea
in the first place, it brings some weird behaviours with it.
This is mostly due to the fact, that the same injection context
is shared for all injection requests which are happening on the
same thread. This causes bindings which were installed to an
injector which is downstream of the current contextual injector
to not have all the bindings present which might be required in
order to inject something into a method.
This is most noticeable when during member injection a second
injector is created which should inject a specific class, with
bindings which are only present in that specific injector. The
thread-local context will force the binding into using a
different injector, which causes exceptions when looking up
parameter instances.
With this commit, the context is no longer bound to an injector
and the injector to which a binding was installed will be used.
This might cause some behaviour changes, as bindings which are
installed downstream of the binding-associated injector are no
longer found, but this is an unintentional usage of the current
implementation anyway.