fix: prevent multiple member injector due to races
there is a possible race between two calls to Injector#memberInjector as the method is not using computeIfAbsent (due to a locking issue in old java versions), but rather a get and put. However, if two threads observe the "not cached" state concurrently there are two member injectors created and returned. Now, by using putIfAbsent instead, one of them will observe the prior insertion and use the previously inserted member injector instead.